ignite-5732: code clean up and refactoring
